AC-392:Download patient option not switching instantly
